script: Move task related files to `script/tasks/`
Maintain the `task_` prefixes, since files like `manager.rs`
are too generic. We did something similar recently with
`audio/node.rs` -> `audio/audio_node.rs`.

Testing: it compiles
