+++ components/script/dom/window/window.rs
@@ -200,8 +200,8 @@ use crate::realms::enter_auto_realm;
 use crate::script_runtime::Runtime;
 use crate::script_thread::ScriptThread;
 use crate::script_window_proxies::ScriptWindowProxies;
-use crate::task_manager::TaskManager;
-use crate::task_source::SendableTaskSource;
+use crate::tasks::task_manager::TaskManager;
+use crate::tasks::task_source::SendableTaskSource;
 use crate::timers::{IsInterval, OneshotTimers, TimerCallback};
 use crate::unminify::unminified_path;
 use crate::webdriver_handlers::{find_node_by_unique_id_in_document, jsval_to_webdriver};
