+++ components/script/tasks/mod.rs
@@ -1,0 +1,10 @@
+/* This Source Code Form is subject to the terms of the Mozilla Public
+ * License, v. 2.0. If a copy of the MPL was not distributed with this
+ * file, You can obtain one at https://mozilla.org/MPL/2.0/. */
+
+#[macro_use]
+pub(crate) mod task;
+
+pub(crate) mod task_manager;
+pub(crate) mod task_queue;
+pub(crate) mod task_source;
