+++ components/script/tasks/task_queue.rs
@@ -16,8 +16,8 @@ use strum::VariantArray;
 
 use crate::dom::worker::TrustedWorkerAddress;
 use crate::script_runtime::ScriptThreadEventCategory;
-use crate::task::TaskBox;
-use crate::task_source::TaskSourceName;
+use crate::tasks::task::TaskBox;
+use crate::tasks::task_source::TaskSourceName;
 
 #[derive(MallocSizeOf)]
 pub(crate) struct QueuedTask {
