+++ components/script/timers.rs
@@ -45,7 +45,7 @@ use crate::dom::xmlhttprequest::XHRTimeoutCallback;
 use crate::modules::script_module::{ScriptFetchOptions, module_script_from_reference_private};
 use crate::script_runtime::IntroductionType;
 use crate::script_thread::ScriptThread;
-use crate::task_source::SendableTaskSource;
+use crate::tasks::task_source::SendableTaskSource;
 
 type TimerKey = i32;
 type RunStepsDeadline = Instant;
